valorantClips/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
Django==4.2.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
mysqlclient==2.2.0
numpy==1.25.2
Pillow==10.0.0
proglog==0.1.10
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
views==0.3
